+++ app/src/main/java/com/dskja/betterstreamflix/providers/StreamingCommunityProvider.kt
@@ -52,8 +52,8 @@ class StreamingCommunityProvider(private val _language: String? = null) : Provid
     private val TAG: String
         get() = "SCProviderDebug[$LANG]"
 
-    private val DEFAULT_DOMAIN: String = "streamingunity.cc"
-    private val BLOCKED_DOMAINS = setOf("streamingcommunityz.green", "streamingunity.club", "streamingunity.bike", "streamingcommunityz.buzz")
+    private val DEFAULT_DOMAIN: String = "streamingunity.win"
+    private val BLOCKED_DOMAINS = setOf("streamingunity.cc", "streamingcommunityz.green", "streamingunity.club", "streamingunity.bike", "streamingcommunityz.buzz")
     override val baseUrl = DEFAULT_DOMAIN
     private var _domain: String? = null
     private var domain: String