+++ app/src/main/java/com/dskja/betterstreamflix/providers/UnJourUnFilmProvider.kt
@@ -50,7 +50,7 @@ object UnJourUnFilmProvider : Provider, ProviderPortalUrl, ProviderConfigUrl {
             return cachePortalURL.ifEmpty { field }
         }
 
-    override val defaultBaseUrl: String = "https://1jour1film0126b.site/"
+    override val defaultBaseUrl: String = "https://1jour1film0826b.website/"
     override val baseUrl: String = defaultBaseUrl
         get() {
             val cacheURL = UserPreferences.getProviderCache(this, UserPreferences.PROVIDER_URL)